requirements_new.txt
absl-py==2.1.0
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
ase==3.23.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
configparser==7.1.0
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.55.3
fqdn==1.5.1
gdown==3.15.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-api-core==2.24.0
google-api-python-client==2.149.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.66.0
googledrivedownloader==1.1.0
gql==0.2.0
graphql-core==1.1
grpcio==1.69.0
h11==0.14.0
h5py==3.2.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hydra-core==0.11.3
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
llvmlite==0.33.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.4.2
matplotlib-inline==0.1.7
mistune==3.1.0
nbclient==0.10.1
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.50.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
omegaconf==1.4.1
open3d==0.10.0.0
overrides==7.7.0
packaging==24.2
pandas==1.4.4
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.1.0
#pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plyfile==0.7.4
pointops2==0.0.0
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-louvain==0.16
pytorch-metric-learning==0.9.99
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==26.2.0
rdflib==7.1.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
rsa==4.9
scikit-image==0.16.2
scikit-learn==1.3.2
scipy==1.6.3
Send2Trash==1.8.3
sentry-sdk==2.19.2
SharedArray==3.2.1
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
subprocess32==3.5.4
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.5
termcolor==1.1.0
terminado==0.18.1
threadpoolctl==3.5.0
timm==0.4.9
tinycss2==1.2.1
tomli==2.2.1
